+++ api/package.json
@@ -3,10 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-
   "scripts": {
-    "start": "nodemon -L",
-    "railway":"node ./index.js",
+    "start": "nodemon index.js",
+    "railway": "node ./index.js",
     "test": "mocha -w ./tests/**/*.spec.js"
   },
   "author": "Henry",
@@ -30,7 +29,7 @@
   "devDependencies": {
     "chai": "^4.2.0",
     "mocha": "^8.2.1",
-    "nodemon": "^2.0.6",
+    "nodemon": "^2.0.22",
     "supertest": "^6.0.1",
     "supertest-session": "^4.1.0"
   }
