+++ api/src/controllers/index.js
@@ -1,0 +1,5 @@
+const { catchedAsync } = require("../utils");
+
+module.exports = {
+  getRecipes: catchedAsync(require("./getRecipes")),
+};
