+++ api/src/data/index.js
@@ -1,0 +1,5 @@
+const fetchRecipes = require("./fetchRecipes")
+
+module.exports = {
+  list: async () => await fetchRecipes(),
+}
