+++ api/src/middlewares/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  getRecipesValidation: require('./getRecipesValidation'),
+}
