+++ api/src/utils/errors/index.js
@@ -1,0 +1,8 @@
+class ClientError extends Error {
+    constructor(message, statusCode = 400) {
+        super(message);
+        this.statusCode = statusCode;
+    }
+} 
+
+module.exports = {ClientError};
