+++ api/src/utils/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  response: require('./response'),
+  catchedAsync: require('./catchedAsync'),
+  errors: require('./errors'),
+}
