+++ api/src/utils/response.js
@@ -1,0 +1,14 @@
+module.exports = (res, statusCode, data) => {
+  res.status(statusCode).json({
+  error: false,
+  data: data,
+  });
+}
+
+
+// module.exports = (res, statusCode, data) => {
+//   res.status(statusCode).json({
+//   error: false,
+//   data,
+//   });
+// }
